+++ R/pixelatorR-package.R
@@ -159,7 +159,6 @@
 #' @importFrom Matrix colSums
 #' @importFrom Matrix rowSums
 #' @importFrom methods as
-#' @importFrom methods .hasSlot
 #' @importFrom methods callNextMethod
 #' @importFrom methods getMethod
 #' @importFrom methods initialize